+++ frontend/src/pages/firebase.js
@@ -1,7 +1,7 @@
 import { initializeApp } from 'firebase/app';
 import { getFirestore } from 'firebase/firestore';
+import { getAuth } from 'firebase/auth';  
 
-// Your Firebase configuration
 const firebaseConfig = {  
     apiKey: import.meta.env.VITE_FIREBASE_API_KEY,  
     authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,  
@@ -12,10 +12,10 @@ const firebaseConfig = {
     measurementId: import.meta.env.VITE_FIREBASE_MEASUREMENT_ID,  
   }; 
 
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
 
-// Get a reference to the Firestore database
+const app = initializeApp(firebaseConfig);
+const auth = getAuth(app);  
+
 const db = getFirestore(app);
 
-export { db }; // Export the db variable
+export { db,auth };
